requirements.txt
requests
PyYAML
python-dateutil
feedparser